Change the output dtype of `find_center_vo_cupy`

+++ httomolib/recon/rotation.py
@@ -111,7 +111,7 @@ def _find_center_vo_gpu(sino, ind, smin, smax, srad, step, ratio, drop):
         init_cen = _search_coarse(_sino_cs, smin, smax, ratio, drop)
         fine_cen = _search_fine(_sino_fs, srad, step, init_cen, ratio, drop)
 
-    return fine_cen
+    return fine_cen.astype(cp.float32)
 
 
 def _search_coarse(sino, smin, smax, ratio, drop):
